requirements.txt
python-telegram-bot==20.7
httpx==0.26.0
python-dotenv==1.0.0
gspread==6.0.0
google-auth==2.26.2
